how to program the LAMP and beeper outputs
The Z80-MBC3 has two digital outputs on the PCB that can be used as a status or warning indicator to the user. The LED labeled "LAMP" not only lights up when the USER button is pressed, but can also be turned on and off per instruction in a Z80 application.

And there's the beeper that issues a short beep when a <BELL> character is sent to the terminal. A simple statement in BASIC demonstrates this with : PRINT CHR$(7).
But it is also possible to beep with variable duration. How this can be programmed is described in this Application Note.