How to use the onboard GPIO or externals via the I2C bus
This application note describes the addressing method of the onboard PCF8574 circuit for general purpose input output (GPIO). And the general addressing scheme to use other peripherals connected to the I2C port of the Z80-MB3. You can add external PCF8574 circuits to expand the number of I/O pins, replace the onboard PCF8574 with an external MCP23017, and/or expand the input output functions with other I2C modules.